Fix calls to management canister
This fixes an issue in the RequestID computation where an empty slice
canister ID would be considered as "canister_id not set". This was not
quite correct, because the management canister ID is the empty slice
(encoded as "aaaaa-aa").

This changes the empty check for a `nil` check instead. Usages of the
effective canister ID vs. canister ID are also clarified.